backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic[email]
bcrypt
scikit-learn
pandas
numpy